Code cleanup

+++ src/GlobalPlatform.NET/Extensions/ByteExtensions.cs
@@ -5,7 +5,7 @@ namespace GlobalPlatform.NET.Extensions
     internal static class ByteExtensions
     {
         /// <summary>
-        /// Adds a range of bytes to the collection, prefixed by a single byte denoting the range's length.
+        /// Adds a range of bytes to the collection, prefixed by a single byte denoting the range's length. 
         /// </summary>
         /// <param name="bytes"></param>
         /// <param name="range"></param>
@@ -21,14 +21,14 @@ namespace GlobalPlatform.NET.Extensions
         }
 
         /// <summary>
-        /// Adds a tag to the collection, followed by the length of the data, followed by the data itself.
+        /// Adds a tag to the collection, followed by the length of the data, followed by the data itself. 
         /// </summary>
         /// <param name="bytes"></param>
         /// <param name="range"></param>
         /// <returns></returns>
         public static byte AddTag(this ICollection<byte> bytes, byte tag, params byte[] data)
         {
-            bytes.Add((byte)tag);
+            bytes.Add(tag);
 
             byte length = data.LengthChecked();
 
@@ -39,7 +39,7 @@ namespace GlobalPlatform.NET.Extensions
         }
 
         /// <summary>
-        /// Returns the length of the array, as a checked byte.
+        /// Returns the length of the array, as a checked byte. 
         /// </summary>
         /// <param name="array"></param>
         /// <returns></returns>
